refactor: refactor all the data feeder
Split csv, sqlite, database and front importer

Solves #14
